requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
httpx
python-dotenv
pydantic
pydantic-settings
cachetools
python-multipart
tenacity